# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KasiDash & BuildForge API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Customers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomersPage"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a single customer with order history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDetail"
...
  /analytics/revenue:
    get:
      operationId: getRevenueAnalytics
      tags: [analytics]
      summary: Get revenue analytics data
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d, 12m]
      responses:
        "200":
          description: Revenue analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueAnalytics"
  /analytics/orders:
    get:
      operationId: getOrderAnalytics
      tags: [analytics]
      summary: Get order analytics data
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d, 12m]
      responses:
        "200":
          description: Order analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderAnalytics"
  /analytics/customers:
    get:
      operationId: getCustomerAnalytics
      tags: [analytics]
      summary: Get customer growth analytics
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d, 12m]
      responses:
        "200":
          description: Customer analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerAnalytics"
...
  /buildforge/systems:
    get:
      operationId: listSystems
      tags: [buildforge]
      summary: List all built systems
      responses:
        "200":
          description: Systems list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BFSystem"
    post:
      operationId: createSystem
      tags: [buildforge]
      summary: Create a new system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BFSystemInput"
      responses:
        "201":
          description: Created system
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BFSystem"
  /buildforge/systems/{id}:
    get:
      operationId: getSystem
      tags: [buildforge]
      summary: Get a system with its modules
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: System detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BFSystemDetail"
    patch:
      operationId: updateSystem
      tags: [buildforge]
      summary: Update a system
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BFSystemUpdate"
      responses:
        "200":
          description: Updated system
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BFSystem"
    delete:
      operationId: deleteSystem
      tags: [buildforge]
      summary: Delete a system
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /buildforge/systems/{id}/modules:
    post:
      operationId: addModule
      tags: [buildforge]
      summary: Add a module to a system
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BFModuleInput"
      responses:
        "201":
          description: Created module
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BFModule"
  /buildforge/modules/{id}:
    patch:
      operationId: updateModule
      tags: [buildforge]
      summary: Update a module (reorder, rename, config)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BFModuleUpdate"
      responses:
        "200":
          description: Updated module
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BFModule"
    delete:
      operationId: deleteModule
      tags: [buildforge]
      summary: Delete a module
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    RevenueDataPoint:
      type: object
      properties:
        date:
          type: string
        revenue:
          type: number
        orders:
          type: integer
      required: [date, revenue, orders]
...
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        customerName:
          type: string
        customerEmail:
          type: string
        status:
          type: string
          enum: [pending, paid, fulfilled, cancelled, refunded]
        total:
          type: number
        itemCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, orderNumber, customerName, customerEmail, status, total, itemCount, createdAt, updatedAt]
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        location:
          type: ["string", "null"]
        totalOrders:
          type: integer
        totalSpent:
          type: number
        lifetimeValue:
          type: number
        isRepeat:
          type: boolean
        createdAt:
          type: string
      required: [id, name, email, totalOrders, totalSpent, lifetimeValue, isRepeat, createdAt]
    CustomerDetail:
      type: object
      properties:
        customer:
          $ref: "#/components/schemas/Customer"
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
      required: [customer, orders]
    CustomersPage:
      type: object
      properties:
        customers:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [customers, total, page, limit]
...
    RevenueAnalytics:
      type: object
      properties:
        dataPoints:
          type: array
          items:
            $ref: "#/components/schemas/RevenueDataPoint"
        totalRevenue:
          type: number
        growth:
          type: number
      required: [dataPoints, totalRevenue, growth]
    OrderAnalytics:
      type: object
      properties:
        dataPoints:
          type: array
          items:
            $ref: "#/components/schemas/RevenueDataPoint"
        totalOrders:
          type: integer
        growth:
          type: number
      required: [dataPoints, totalOrders, growth]
    CustomerAnalytics:
      type: object
      properties:
        dataPoints:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              newCustomers:
                type: integer
              totalCustomers:
                type: integer
            required: [date, newCustomers, totalCustomers]
        totalNewCustomers:
          type: integer
        growth:
          type: number
      required: [dataPoints, totalNewCustomers, growth]
...
    BFSystem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [invoice, employee, crm, booking, delivery, stock, expense, payroll, custom]
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, active, archived]
        moduleCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, type, status, moduleCount, createdAt, updatedAt]
    BFSystemInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [invoice, employee, crm, booking, delivery, stock, expense, payroll, custom]
        description:
          type: string
      required: [name, type]
    BFSystemUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [draft, active, archived]
    BFModule:
      type: object
      properties:
        id:
          type: integer
        systemId:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [table, form, dashboard, report, chart, permissions]
        config:
          type: object
        sortOrder:
          type: integer
        createdAt:
          type: string
      required: [id, systemId, name, type, config, sortOrder, createdAt]
    BFModuleInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [table, form, dashboard, report, chart, permissions]
        config:
          type: object
        sortOrder:
          type: integer
      required: [name, type]
    BFModuleUpdate:
      type: object
      properties:
        name:
          type: string
        config:
          type: object
        sortOrder:
          type: integer
    BFSystemDetail:
      type: object
      properties:
        system:
          $ref: "#/components/schemas/BFSystem"
        modules:
          type: array
          items:
            $ref: "#/components/schemas/BFModule"
      required: [system, modules]
...