# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * KasiDash & BuildForge API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  orderNumber: string;
  customerName: string;
  customerEmail: string;
  status: OrderStatus;
  total: number;
  itemCount: number;
  createdAt: string;
  updatedAt: string;
}